Rimossi refusi float dalla classe Spectrum

+++ inc/htr_spectrum.h
@@ -24,7 +24,7 @@ namespace Spectrometry {
       Spectrum(const std::vector <T> &hist);
 
       template <typename T>
-      Spectrum(const std::vector <T> &hist, const Epoch::DateTime &start, float tm);
+      Spectrum(const std::vector <T> &hist, const Epoch::DateTime &start, double tm);
 
       double binAt(int b1) const;
       std::size_t channels() const;
@@ -38,12 +38,12 @@ namespace Spectrometry {
       Spectrum& rebin(double gain, unsigned seed);
 
     private:
-      std::vector <float> bin;
+      std::vector <double> bin;
       std::size_t canali;
       double mCal;
       double qCal;
       Epoch::DateTime startTime;
-      float dT;
+      double dT;
   };
 
   template <typename T>
@@ -58,7 +58,7 @@ namespace Spectrometry {
   }
 
   template <typename T>
-    Spectrum::Spectrum(const std::vector <T> &hist, const Epoch::DateTime &start, float tm)
+    Spectrum::Spectrum(const std::vector <T> &hist, const Epoch::DateTime &start, double tm)
       : Spectrum(hist)
   {
     startTime = start;
